+++ src/usb/daqi/DaqIUsbBase.cpp
@@ -131,7 +131,7 @@ int DaqIUsbBase::calcStageSize(int epAddr, double rate, int chanCount, int sampl
 	else
 	{
 		double aggRate =  chanCount * rate * sampleSize; // bytes per second
-		long long bufferBytesCount = sampleCount * sampleSize;
+		long long bufferBytesCount = (long long) sampleCount * sampleSize;
 		double stageRate = daqDev().scanTranserIn()->getStageRate();
 		stageSize = (int)(aggRate * stageRate);
